refactor: move test modules from lib.rs into files they exercise
TD-010: Moved 137 unit tests from lib.rs into #[cfg(test)] modules
inside the files they actually exercise:
- parser.rs: system command and classification tests
- intent_local.rs: local intent parsing tests
- commands.rs: validation helper tests
- mention.rs: @mention extraction tests
- intent_types.rs: IntentKind deserialization tests
- intent_llm.rs: IntentResponse deserialization tests

lib.rs is now a clean re-export module with no test code.
